Remove containers folder and rename PlayersList to PlayerList
