+++ client/src/components/PlayerList/PlayerList.scss
@@ -1,6 +1,6 @@
 @import 'styles/variables';
 
-.PlayersList {
+.PlayerList {
   h2 {
     margin-bottom: $spacing-small;
   }
