+++ client/src/components/PlayerList/PlayerList.test.tsx
@@ -1,23 +1,24 @@
 import React from 'react'
 import { render } from '@testing-library/react'
-import PlayersList from './PlayersList'
+import PlayerList from './PlayerList'
+import { Player } from '../../types/Player'
 
-const playersMock = [
-  { name: 'Beth', id: '1' },
-  { name: 'John', id: '2' }
+const playersMock: Array<Player> = [
+  { name: 'Beth', id: '1', score: 0 },
+  { name: 'John', id: '2', score: 0 }
 ]
 
-describe('PlayersList', () => {
+describe('PlayerList', () => {
   it('renders without crashing', () => {
     const { container } = render(
-      <PlayersList players={playersMock} currentPlayerId='2' />
+      <PlayerList players={playersMock} currentPlayerId='2' />
     )
     expect(container).toBeInTheDocument()
   })
 
   it('renders the correct number of players', () => {
     const { getAllByRole } = render(
-      <PlayersList players={playersMock} currentPlayerId='2' />
+      <PlayerList players={playersMock} currentPlayerId='2' />
     )
     const items = getAllByRole('listitem')
     expect(items).toHaveLength(2)
