+++ client/src/components/PlayerList/index.ts
@@ -1,0 +1,2 @@
+export * from './PlayerList'
+export { default } from './PlayerList'
