--- client/src/containers/PlayersList/index.ts
@@ -1,2 +1,0 @@
-export * from './PlayersList'
-export { default } from './PlayersList'
